requirements.txt
fastapi==0.111.0
hydra-core==1.3.2
loguru==0.7.2
omegaconf==2.3.0
pandas==2.2.2
python-dotenv==1.0.1
pytorch_lightning==2.2.5
scikit_learn==1.4.2
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
typer==0.12.3
wandb>=0.12.10
pytest==8.2.2
coverage==7.5.3
pre-commit==3.7.1
ruff==0.4.10
uvicorn==0.30.1
typer==0.12.3
PyYAML==6.0
starlette==0.37.2
kaggle==1.6.14
torch-model-archiver==0.11.0
streamlit==1.36.0
matplotlib==3.9.0
watchdog==4.0.1
seaborn==0.13.2
locust==2.29.1
dvc==3.51.1
dvc-gdrive==3.0.1
dvc_gs==3.0.1
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
evidently==0.4.30
prometheus-fastapi-instrumentator==7.0.0
mkdocs-material==9.5.28
